transformations.py + updates
moving transformation functions over to a new document, transformations.py
deleting duplicate definitions of corpusdir, newarticledir
updating import statements for new file structure
